types: preserve UnspecifiedFsp through Datum.decimal round-trip
Shrinking Datum.decimal from uint16 to uint8 in 2a3c81e006 dropped the
int8 sign-extending cast in GetMysqlDuration, so Fsp=-1 (UnspecifiedFsp)
came back as 255 and crashed Duration.formatFrac with a slice-bounds
panic (TestDateArithFuncs).

Restore the int8 round-trip so the -1 sentinel survives the uint storage.

+++ pkg/types/datum.go
@@ -395,9 +395,11 @@ func (d *Datum) SetMysqlDecimal(b *MyDecimal) {
 	d.x = b
 }
 
-// GetMysqlDuration gets Duration value
+// GetMysqlDuration gets Duration value. The int8 round-trip preserves
+// UnspecifiedFsp (-1), which callers use as a "no explicit fsp" sentinel;
+// a plain int(d.decimal) would surface it as 255 and crash formatters.
 func (d *Datum) GetMysqlDuration() Duration {
-	return Duration{Duration: time.Duration(d.i), Fsp: int(d.decimal)}
+	return Duration{Duration: time.Duration(d.i), Fsp: int(int8(d.decimal))}
 }
 
 // SetMysqlDuration sets Duration value
